datapath: ipfrag: extend flow key with maximum fragment size
This patch extends flow key with the maximum fragment size which is
non-zero if at least one fragment had DF set. It's a metadata, since
it doesn't occupy IP header. The maximum fragment size is used when
a packet is fragmented in order to preserve the maximum fragment size
of the packet across inbound and outbound physical vports.

Because the maximum fragment size is stored in OVS_CB as part of
a flow key, it's always available at outgoing vports.

Signed-off-by: Takayuki Usui <takayuki@midokura.com>